fix: add missing fields to pinterest_tag mapping (#3814)
* feat: add missing fields on the pinterest_tag mapping

* chore: fix test cases

* chore: remove redundant old implementation

* chore: remove test cases related to pinterest_tag

* chore: fix test cases

* chore: fix lint issue

* chore: fix lint issue again

* chore: revert service.api.test.ts changes
