--- src/v0/destinations/pinterest_tag/data/pinterestCustomConfig.json
@@ -1,38 +1,0 @@
-[
-  {
-    "destKey": "currency",
-    "sourceKeys": "properties.currency",
-    "required": false
-  },
-  {
-    "destKey": "value",
-    "sourceKeys": ["properties.value", "properties.total", "properties.revenue"],
-    "required": false,
-    "metadata": {
-      "type": "toString"
-    }
-  },
-  {
-    "destKey": "num_items",
-    "sourceKeys": "properties.numOfItems",
-    "metadata": {
-      "type": "toInt"
-    },
-    "required": false
-  },
-  {
-    "destKey": "order_id",
-    "sourceKeys": "properties.order_id",
-    "required": false
-  },
-  {
-    "destKey": "search_string",
-    "sourceKeys": "properties.query",
-    "required": false
-  },
-  {
-    "destKey": "opt_out_type",
-    "sourceKeys": "properties.optOutType",
-    "required": false
-  }
-]
