+++ test/__tests__/data/pinterest_tag_output.json
@@ -41,10 +41,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -58,6 +66,7 @@
       "files": {}
     }
   ],
+
   [
     {
       "version": "1",
@@ -96,10 +105,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -210,10 +227,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -277,10 +302,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -649,6 +682,10 @@
             "content_ids": ["507f1f77bcf86cd799439011"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               }
@@ -752,10 +789,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -808,10 +853,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
@@ -888,10 +941,18 @@
             "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
             "contents": [
               {
+                "id": "507f1f77bcf86cd799439011",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Monopoly: 3rd Edition",
                 "quantity": 1,
                 "item_price": "19"
               },
               {
+                "id": "505bd76785ebb509fc183733",
+                "item_brand": "testBrand",
+                "item_category": "Games",
+                "item_name": "Uno Card Game",
                 "quantity": 2,
                 "item_price": "3"
               }
