+++ test/__tests__/data/pinterest_tag_router_input.json
@@ -48,6 +48,7 @@
           "optOutType": "LDP",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -58,6 +59,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -170,6 +172,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -180,6 +183,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -395,6 +399,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -405,6 +410,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -517,6 +523,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -527,6 +534,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -639,6 +647,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -649,6 +658,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -866,6 +876,7 @@
           "optOutType": "LDP",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -876,6 +887,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -990,6 +1002,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -1000,6 +1013,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
@@ -1114,6 +1128,7 @@
           "requestIP": "123.0.0.0",
           "products": [
             {
+              "brand": "testBrand",
               "sku": "45790-32",
               "url": "https://www.example.com/product/path",
               "name": "Monopoly: 3rd Edition",
@@ -1124,6 +1139,7 @@
               "product_id": "507f1f77bcf86cd799439011"
             },
             {
+              "brand": "testBrand",
               "sku": "46493-32",
               "name": "Uno Card Game",
               "price": 3,
