+++ test/__tests__/data/pinterest_tag_router_output.json
@@ -43,10 +43,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -83,10 +91,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -122,10 +138,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -196,10 +220,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -235,10 +267,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -465,10 +505,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -504,10 +552,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -542,10 +598,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
@@ -580,10 +644,18 @@
                   "content_ids": ["507f1f77bcf86cd799439011", "505bd76785ebb509fc183733"],
                   "contents": [
                     {
+                      "id": "507f1f77bcf86cd799439011",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Monopoly: 3rd Edition",
                       "quantity": 1,
                       "item_price": "19"
                     },
                     {
+                      "id": "505bd76785ebb509fc183733",
+                      "item_brand": "testBrand",
+                      "item_category": "Games",
+                      "item_name": "Uno Card Game",
                       "quantity": 2,
                       "item_price": "3"
                     }
