+++ test/apitests/data_scenarios/destination/proc/batch_input_multiplex.json
@@ -78,6 +78,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
@@ -254,6 +255,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
