+++ test/apitests/data_scenarios/destination/proc/multiplex_failure.json
@@ -77,6 +77,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
@@ -206,13 +207,13 @@
         "destinationDefinitionId": ""
       },
       "statusCode": 400,
-      "error": "Missing required value from [\"traits.action_source\",\"context.traits.action_source\",\"traits.actionSource\",\"context.traits.actionSource\",\"properties.action_source\",\"properties.actionSource\",\"channel\"]",
+      "error": "action_source is required: Workflow: procWorkflow, Step: validateCommonFields, ChildStep: undefined, OriginalError: action_source is required",
       "statTags": {
         "errorCategory": "dataValidation",
         "errorType": "instrumentation",
         "destType": "PINTEREST_TAG",
         "module": "destination",
-        "implementation": "native",
+        "implementation": "cdkV2",
         "feature": "processor",
         "destinationId": "2JIqVoWNvSOHa9ppKOqSo4hPuc0",
         "workspaceId": "27O0bhB6p5ehfOWeeZlOSsSDTLg"
