+++ test/apitests/data_scenarios/destination/proc/multiplex_partial_failure.json
@@ -78,6 +78,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
@@ -254,6 +255,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
@@ -543,13 +545,13 @@
         "destinationDefinitionId": ""
       },
       "statusCode": 400,
-      "error": "Action source must be one of app_android, app_ios, web, offline",
+      "error": "Action source must be one of app_android, app_ios, web, offline: Workflow: procWorkflow, Step: validateCommonFields, ChildStep: undefined, OriginalError: Action source must be one of app_android, app_ios, web, offline",
       "statTags": {
         "errorCategory": "dataValidation",
         "errorType": "instrumentation",
         "destType": "PINTEREST_TAG",
         "module": "destination",
-        "implementation": "native",
+        "implementation": "cdkV2",
         "feature": "processor",
         "destinationId": "2JIqVoWNvSOHa9ppKOqSo4hPuc0",
         "workspaceId": "27O0bhB6p5ehfOWeeZlOSsSDTLg"
