+++ test/apitests/data_scenarios/destination/proc/multiplex_success.json
@@ -78,6 +78,7 @@
           "DisplayName": "Pinterest Tag",
           "Config": {
             "cdkV2TestThreshold": 1,
+            "cdkV2Enabled": true,
             "destConfig": {
               "defaultConfig": [
                 "tagId",
