+++ test/integrations/destinations/pinterest_tag/processor/data.ts
@@ -26,13 +26,21 @@ export const data = [
                   lastname: 'Rudderlabs',
                   firstName: 'Test',
                   address: { city: 'Kolkata', state: 'WB', zip: '700114', country: 'IN' },
+                  partnerId: '12345678',
                 },
                 device: { advertisingId: 'abc123' },
                 library: { name: 'rudder-sdk-ruby-sync', version: '1.0.6' },
+                network: {
+                  carrier: 'VI',
+                  wifi: true,
+                },
               },
               messageId: '7208bbb6-2c4e-45bb-bf5b-ad426f3593e9',
               timestamp: '2020-08-14T05:30:30.118Z',
               properties: {
+                contentBrand: 'LV',
+                partnerName: 'testPartner',
+                contentName: 'testContent',
                 tax: 2,
                 total: 27.5,
                 coupon: 'hasbros',
@@ -106,55 +114,77 @@ export const data = [
           {
             metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
             output: {
-              version: '1',
-              type: 'REST',
-              method: 'POST',
-              endpoint: 'https://ct.pinterest.com/events/v3',
-              headers: { 'Content-Type': 'application/json' },
-              params: { test: true },
               body: {
+                FORM: {},
                 JSON: {
                   action_source: 'web',
+                  advertiser_id: '429047995',
+                  app_id: '429047995',
+                  custom_data: {
+                    content_brand: 'LV',
+                    content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
+                    content_name: 'testContent',
+                    contents: [
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        item_price: '19',
+                        quantity: 1,
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        item_price: '3',
+                        quantity: 2,
+                      },
+                    ],
+                    currency: 'USD',
+                    num_items: 3,
+                    opt_out_type: 'LDP',
+                    order_id: '50314b8e9bcf000000000000',
+                    value: '27.5',
+                  },
+                  device_carrier: 'VI',
+                  event_id: '7208bbb6-2c4e-45bb-bf5b-ad426f3593e9',
                   event_name: 'watch_video',
                   event_time: 1597383030,
-                  event_id: '7208bbb6-2c4e-45bb-bf5b-ad426f3593e9',
-                  app_id: '429047995',
-                  advertiser_id: '429047995',
+                  partner_name: 'testPartner',
                   user_data: {
-                    em: ['48ddb93f0b30c475423fe177832912c5bcdce3cc72872f8051627967ef278e08'],
-                    ph: ['d164bbe036663cb5c96835e9ccc6501e9a521127ea62f6359744928ba932413b'],
-                    ln: ['dcf000c2386fb76d22cefc0d118a8511bb75999019cd373df52044bccd1bd251'],
-                    fn: ['9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08'],
-                    ct: ['6689106ca7922c30b2fd2c175c85bc7fc2d52cc4941bdd7bb622c6cdc6284a85'],
-                    st: ['3b45022ab36728cdae12e709e945bba267c50ee8a91e6e4388539a8e03a3fdcd'],
-                    zp: ['1a4292e00780e18d00e76fde9850aee5344e939ba593333cd5e4b4aa2cd33b0c'],
+                    click_id: 'dummy_clickId',
+                    client_user_agent: 'chrome',
                     country: ['582967534d0f909d196b97f9e6921342777aea87b46fa52df165389db1fb8ccf'],
+                    ct: ['6689106ca7922c30b2fd2c175c85bc7fc2d52cc4941bdd7bb622c6cdc6284a85'],
+                    em: ['48ddb93f0b30c475423fe177832912c5bcdce3cc72872f8051627967ef278e08'],
+                    fn: ['9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08'],
+                    ge: ['1b16b1df538ba12dc3f97edbb85caa7050d46c148134290feba80f8236c83db9'],
                     hashed_maids: [
                       '6ca13d52ca70c883e0f0bb101e425a89e8624de51db2d2392593af6a84118090',
                     ],
-                    client_user_agent: 'chrome',
-                    click_id: 'dummy_clickId',
-                    ge: ['1b16b1df538ba12dc3f97edbb85caa7050d46c148134290feba80f8236c83db9'],
+                    ln: ['dcf000c2386fb76d22cefc0d118a8511bb75999019cd373df52044bccd1bd251'],
+                    partner_id: '12345678',
+                    ph: ['d164bbe036663cb5c96835e9ccc6501e9a521127ea62f6359744928ba932413b'],
+                    st: ['3b45022ab36728cdae12e709e945bba267c50ee8a91e6e4388539a8e03a3fdcd'],
+                    zp: ['1a4292e00780e18d00e76fde9850aee5344e939ba593333cd5e4b4aa2cd33b0c'],
                   },
-                  custom_data: {
-                    currency: 'USD',
-                    value: '27.5',
-                    order_id: '50314b8e9bcf000000000000',
-                    opt_out_type: 'LDP',
-                    num_items: 3,
-                    content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
-                    contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
-                    ],
-                  },
+                  wifi: true,
                 },
                 JSON_ARRAY: {},
                 XML: {},
-                FORM: {},
               },
+              endpoint: 'https://ct.pinterest.com/events/v3',
               files: {},
+              headers: {
+                'Content-Type': 'application/json',
+              },
+              method: 'POST',
+              params: {
+                test: true,
+              },
+              type: 'REST',
               userId: '',
+              version: '1',
             },
             statusCode: 200,
           },
@@ -302,8 +332,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -835,8 +877,20 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -1357,8 +1411,20 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -2486,7 +2552,15 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     num_items: 1,
                     content_ids: ['507f1f77bcf86cd799439011'],
-                    contents: [{ quantity: 1, item_price: '19' }],
+                    contents: [
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                    ],
                   },
                 },
                 JSON_ARRAY: {},
@@ -2785,8 +2859,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -2951,8 +3037,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -2980,7 +3078,7 @@ export const data = [
         body: [
           {
             description:
-              'Custom event with v5 Api version, with unhashed User Data and the values are an array of strings',
+              'Custom event with v5 Api version, with unhashed User Data and the values are an of strings',
             message: {
               type: 'track',
               event: 'random',
@@ -3147,8 +3245,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
