+++ test/integrations/destinations/pinterest_tag/router/data.ts
@@ -674,8 +674,20 @@ export const data = [
                           opt_out_type: 'LDP',
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -710,8 +722,20 @@ export const data = [
                           opt_out_type: 'LDP',
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -745,8 +769,20 @@ export const data = [
                           num_items: 3,
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -812,8 +848,20 @@ export const data = [
                           num_items: 3,
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -847,8 +895,20 @@ export const data = [
                           num_items: 3,
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -981,7 +1041,7 @@ export const data = [
   },
   {
     destType: 'pinterest_tag',
-    description: 'Test 0',
+    description: 'Test 1',
     feature: 'router',
     module: 'destination',
     version: 'v0',
@@ -1328,8 +1388,20 @@ export const data = [
                           opt_out_type: 'LDP',
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -1363,8 +1435,20 @@ export const data = [
                           opt_out_type: 'LDP',
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -1397,8 +1481,20 @@ export const data = [
                           num_items: 3,
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -1431,8 +1527,20 @@ export const data = [
                           num_items: 3,
                           content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                           contents: [
-                            { quantity: 1, item_price: '19' },
-                            { quantity: 2, item_price: '3' },
+                            {
+                              id: '507f1f77bcf86cd799439011',
+                              item_category: 'Games',
+                              item_name: 'Monopoly: 3rd Edition',
+                              quantity: 1,
+                              item_price: '19',
+                            },
+                            {
+                              id: '505bd76785ebb509fc183733',
+                              item_category: 'Games',
+                              item_name: 'Uno Card Game',
+                              quantity: 2,
+                              item_price: '3',
+                            },
                           ],
                         },
                       },
@@ -1488,7 +1596,7 @@ export const data = [
   },
   {
     destType: 'pinterest_tag',
-    description: 'Test 0',
+    description: 'Test 2',
     feature: 'router',
     module: 'destination',
     version: 'v0',
