+++ test/integrations/destinations/pinterest_tag/step/data.ts
@@ -75,6 +75,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -90,6 +91,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -101,6 +103,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -139,8 +144,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -232,6 +249,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -247,6 +265,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -258,6 +277,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -294,8 +316,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -368,6 +402,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -383,6 +418,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -394,6 +430,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -519,6 +558,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -534,6 +574,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -545,13 +586,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Advertiser Id not found. Aborting',
+            error:
+              'Advertiser Id not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Advertiser Id not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -633,6 +678,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -648,6 +694,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -660,13 +707,16 @@ export const data = [
         body: [
           {
             error:
-              'It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent',
+              'It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent: Workflow: procWorkflow, Step: validateUserFields, ChildStep: undefined, OriginalError: It is required at least one of em, hashed_maids or pair of client_ip_address and client_user_agent',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -751,6 +801,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -766,6 +817,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -777,6 +829,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -813,8 +868,20 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -907,6 +974,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -922,6 +990,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -933,13 +1002,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Advertiser Id not found. Aborting',
+            error:
+              'Advertiser Id not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Advertiser Id not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1023,6 +1096,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1038,6 +1112,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1049,13 +1124,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'message type group is not supported',
+            error:
+              'message type group is not supported: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: message type group is not supported',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1140,6 +1219,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1155,6 +1235,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1166,13 +1247,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Action source must be one of app_android, app_ios, web, offline',
+            error:
+              'Action source must be one of app_android, app_ios, web, offline: Workflow: procWorkflow, Step: validateCommonFields, ChildStep: undefined, OriginalError: Action source must be one of app_android, app_ios, web, offline',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'instrumentation',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -1257,6 +1342,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1273,6 +1359,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1284,6 +1371,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1320,8 +1410,20 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -1378,6 +1480,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1393,6 +1496,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1404,6 +1508,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1488,6 +1595,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1503,6 +1611,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1514,6 +1623,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1598,6 +1710,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1613,6 +1726,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1624,6 +1738,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1709,6 +1826,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1724,6 +1842,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1735,6 +1854,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1821,6 +1943,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1834,6 +1957,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1845,6 +1969,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -1933,6 +2060,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -1946,6 +2074,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -1957,6 +2086,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2096,6 +2228,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2114,6 +2247,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2125,6 +2259,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               body: {
                 JSON: {
@@ -2216,6 +2353,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2232,6 +2370,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2243,6 +2382,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2360,6 +2502,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2376,6 +2519,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2387,6 +2531,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2422,7 +2569,15 @@ export const data = [
                     order_id: '50314b8e9bcf000000000000',
                     num_items: 1,
                     content_ids: ['507f1f77bcf86cd799439011'],
-                    contents: [{ quantity: 1, item_price: '19' }],
+                    contents: [
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                    ],
                   },
                 },
                 JSON_ARRAY: {},
@@ -2491,6 +2646,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2506,6 +2662,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2517,6 +2674,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               body: {
                 FORM: {},
@@ -2644,6 +2804,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2662,6 +2823,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2673,6 +2835,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2715,8 +2880,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -2810,6 +2987,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2828,6 +3006,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -2839,6 +3018,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -2878,8 +3060,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -2979,6 +3173,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -2997,6 +3192,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3008,6 +3204,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
@@ -3071,8 +3270,20 @@ export const data = [
                     num_items: 3,
                     content_ids: ['507f1f77bcf86cd799439011', '505bd76785ebb509fc183733'],
                     contents: [
-                      { quantity: 1, item_price: '19' },
-                      { quantity: 2, item_price: '3' },
+                      {
+                        id: '507f1f77bcf86cd799439011',
+                        item_category: 'Games',
+                        item_name: 'Monopoly: 3rd Edition',
+                        quantity: 1,
+                        item_price: '19',
+                      },
+                      {
+                        id: '505bd76785ebb509fc183733',
+                        item_category: 'Games',
+                        item_name: 'Uno Card Game',
+                        quantity: 2,
+                        item_price: '3',
+                      },
                     ],
                   },
                 },
@@ -3129,6 +3340,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3147,6 +3359,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3158,13 +3371,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Ad Account ID not found. Aborting',
+            error:
+              'Ad Account ID not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Ad Account ID not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -3213,6 +3430,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3231,6 +3449,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3242,13 +3461,17 @@ export const data = [
         status: 200,
         body: [
           {
-            error: 'Conversion Token not found. Aborting',
+            error:
+              'Conversion Token not found. Aborting: Workflow: procWorkflow, Step: validateInput, ChildStep: undefined, OriginalError: Conversion Token not found. Aborting',
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             statTags: {
               destType: 'PINTEREST_TAG',
               errorCategory: 'dataValidation',
               errorType: 'configuration',
               feature: 'processor',
-              implementation: 'native',
+              implementation: 'cdkV2',
               module: 'destination',
             },
             statusCode: 400,
@@ -3310,6 +3533,7 @@ export const data = [
             destination: {
               ID: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
               Name: 'PINTEREST_TAG',
+              DestinationDefinition: { Config: { cdkV2Enabled: true } },
               Config: {
                 sendAsTestEvent: false,
                 tagId: '123456789',
@@ -3328,6 +3552,7 @@ export const data = [
               Enabled: true,
               Transformations: [],
             },
+            metadata: { destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq' },
           },
         ],
         method: 'POST',
@@ -3339,6 +3564,9 @@ export const data = [
         status: 200,
         body: [
           {
+            metadata: {
+              destintionId: '1pYpzzvcn7AQ2W9GGIAZSsN6Mfq',
+            },
             output: {
               version: '1',
               type: 'REST',
